--- packages/grafana-i18n/src/types/dates.d.ts
@@ -1,13 +1,0 @@
-import {
-  DurationFormatConstructor,
-  DurationFormatOptions as _DurationFormatOptions,
-  DurationInput as _DurationInput,
-} from '@formatjs/intl-durationformat/src/types';
-
-declare global {
-  namespace Intl {
-    const DurationFormat: DurationFormatConstructor;
-    type DurationFormatOptions = _DurationFormatOptions;
-    type DurationInput = _DurationInput;
-  }
-}
